refactoring day05
