+++ 2019/src/lib.rs
@@ -4,3 +4,4 @@ pub mod day01;
 pub mod day02;
 pub mod day03;
 pub mod day04;
+pub mod day05;
