+++ 2019/tests/day05_test.rs
@@ -1,0 +1,14 @@
+#[cfg(test)]
+mod tests {
+    use aoc2019::day05;
+
+    #[test]
+    fn part1_works() {
+        assert_eq!(9006673, day05::part1());
+    }
+
+    #[test]
+    fn part2_works() {
+        assert_eq!(3629692, day05::part2());
+    }
+}
